Add link to current `libm` project
There is an item on the list about a `libm` implementation in Rust. The `m` project mentioned has been deprecated in favor of a new `libm` project; add link to the new project. It is still a work in progress so it still belongs on this list, there's just a newer project with more progress by now.