Adding a Game Developement section
Game development is a huge industry, also automotive, simulation, architecture industries use tools originally made for game development. Even though Rust is very much capable of changing these industries, the lack of these tools written in Rust or which allow Rust usage makes it so that this should be added to this list in my opinion.
I'm adding this PR after u/ErichDonGubler 's suggestion on a Reddit post of some time ago. I'm finally doing it after so long.